sql: rename systemprivilege -> syntheticprivilege package
minor package rename and reorg

Release note: None
