+++ pkg/sql/BUILD.bazel
@@ -420,6 +420,7 @@ go_library(
         "//pkg/sql/storageparam",
         "//pkg/sql/storageparam/indexstorageparam",
         "//pkg/sql/storageparam/tablestorageparam",
+        "//pkg/sql/syntheticprivilege",
         "//pkg/sql/types",
         "//pkg/sql/vtable",
         "//pkg/storage/enginepb",
