+++ pkg/sql/catalog/BUILD.bazel
@@ -9,10 +9,10 @@ go_library(
         "descriptor.go",
         "descriptor_id_set.go",
         "errors.go",
-        "global_privilege.go",
         "post_derserialization_changes.go",
         "privilege_object.go",
         "schema.go",
+        "synthetic_privilege.go",
         "system_table.go",
         "table_col_map.go",
         "table_col_set.go",
@@ -27,7 +27,6 @@ go_library(
         "//pkg/keys",
         "//pkg/kv",
         "//pkg/roachpb",
-        "//pkg/security/username",
         "//pkg/server/telemetry",
         "//pkg/sql/catalog/catpb",
         "//pkg/sql/catalog/descpb",
@@ -35,10 +34,8 @@ go_library(
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/privilege",
         "//pkg/sql/schemachanger/scpb",
-        "//pkg/sql/sem/catconstants",
         "//pkg/sql/sem/eval",
         "//pkg/sql/sem/tree",
-        "//pkg/sql/sessiondata",
         "//pkg/sql/types",
         "//pkg/util",
         "//pkg/util/hlc",
