+++ pkg/sql/catalog/synthetic_privilege.go
@@ -1,0 +1,21 @@
+// Copyright 2022 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package catalog
+
+import "github.com/cockroachdb/cockroach/pkg/sql/privilege"
+
+// SyntheticPrivilegeObject represents an object that has its privileges stored
+// in system.privileges.
+type SyntheticPrivilegeObject interface {
+	PrivilegeObject
+	ToString() string
+	PrivilegeObjectType() privilege.ObjectType
+}
